refactor: room.service.ts 에서 User 인스턴스를 생성하고 있어서 단일 책임 원칙을 위해 로직 분리
user.factory.ts 에서 랜덤 User 인스턴스 생성하고, room.service 에 user factory 주입
