+++ backend/chatServer/src/user/user.interface.ts
